refactor to use util api calls
