+++ Components/MapComponent.jsx
@@ -5,7 +5,7 @@ import * as Location from "expo-location";
 import MapViewDirections from "react-native-maps-directions";
 import Example from "./BottomSheet";
 import mapStyle from "../assets/MapStyle";
-import { getCurrentQuest, getQuests, getUser } from "../utils/users";
+import { getSingularQuest, getQuests, getUser } from "../utils/api";
 
 const Map = () => {
   const [currentLocation, setCurrentLocation] = useState(null);
@@ -30,7 +30,7 @@ const Map = () => {
         const user = await getUser();
         setCurrentQuest(user.currentQuest);
 
-        getCurrentQuest(setCurrentQuest, currentQuest, setQuestDestination);
+        getSingularQuest(setCurrentQuest, currentQuest, setQuestDestination);
 
         const locationSubscription = await Location.watchPositionAsync(
           {
